// FoodDiary.Web.Client/src/app/shared/api/sdk/generated/model/hydration-entry-http-response.ts
/**
 * FoodDiary User API
 *
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

export interface HydrationEntryHttpResponse {
    amountMl?: number;
    id?: string;
    timestampUtc?: string;
}